Moved `string_to_coq` for better decoupling
